separator ts and tests
