--- src/components/Separator/Separator.mdx
@@ -1,10 +1,0 @@
-import { Playground, Props } from 'docz'
-import Separator from '@components/Separator'
-
-# Separator
-
-<Props of={Separator} />
-
-<Playground>
-  <Separator>Lorem ipsum</Separator>
-</Playground>
