+++ src/components/Separator/Separator.test.tsx
@@ -1,0 +1,9 @@
+import { render } from '@testing-library/react'
+
+import { Separator } from './Separator'
+
+describe('Separator tests', () => {
+  it('should renders properly', () => {
+    render(<Separator>Lorem ipsum</Separator>)
+  })
+})
